requirements.txt
Django==5.1.1
djangorestframework==3.15.2
django-cors-headers==4.4.0
psycopg2-binary==2.9.9
django-environ==0.11.2
gunicorn==23.0.0
dj-database-url==2.2.0
whitenoise==6.7.0